requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.1.1
iniconfig==2.0.0
openpyxl==3.1.2
packaging==23.0
pluggy==1.0.0
py==1.11.0
pytest==7.3.0
pytest-html==3.2.0
pytest-metadata==2.0.4
requests==2.28.2
tomli==2.0.1
urllib3==1.26.15
PyYAML==6.0
utils==1.0.1
PyMySQL==1.0.2
geoip2~=4.7.0